Move granularity tests to numericRestrictionsTests
